Clarify configuration and SNMP OID guidance
Improve novice documentation, explain bits and bytes, and document the lightweight MIB-to-OID abstraction.

+++ src/fixed-speeds/main.cpp
@@ -5,7 +5,7 @@ void getSNMP();
 void resetDelayTimer();
 void stopFastPolling();
 bool isValidPoll();
-float calculateBandwidth(uint32_t current, uint32_t last, uint32_t speed, uint32_t currentTime, uint32_t lastTime);
+float calculateBandwidth(uint64_t current, uint64_t last, uint32_t speed, uint32_t currentTime, uint32_t lastTime);
 void updateDisplay();
 
 #include "../../broadbandspeed_FixedSpeeds.ino"
